Godung function

+++ src/main/java/com/yeamgood/godungonline/repository/RoleLoginRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.yeamgood.godungonline.model.RoleLogin;
 
 @Repository("roleloginRepository")
-public interface RoleLoginRepository extends JpaRepository<RoleLogin, Integer>{
+public interface RoleLoginRepository extends JpaRepository<RoleLogin, Long>{
 	RoleLogin findByRole(String role);
 
 }
